synthetic_multi/requirements.txt
pandas>=2.0.0
pyyaml>=6.0
sdv>=1.10.0
numpy>=1.23.0
rich>=13.0.0